Remove `CustomRecordStore` from public API, replace it with internal `ProviderOnlyRecordStore`
