+++ crates/subspace-networking/src/behavior.rs
@@ -1,5 +1,5 @@
-pub(crate) mod custom_record_store;
 pub(crate) mod persistent_parameters;
+pub(crate) mod provider_storage;
 #[cfg(test)]
 mod tests;
