+++ crates/subspace-networking/src/lib.rs
@@ -40,8 +40,8 @@ pub use crate::node::{
     TopicSubscription,
 };
 pub use crate::node_runner::NodeRunner;
-pub use behavior::custom_record_store::{
-    CustomRecordStore, MemoryProviderStorage, ParityDbProviderStorage, ProviderStorage,
+pub use behavior::provider_storage::{
+    MemoryProviderStorage, ParityDbProviderStorage, ProviderStorage,
 };
 pub use create::{create, peer_id, Config, CreationError, RelayMode};
 pub use libp2p;
