+++ crates/subspace-networking/src/node_runner.rs
@@ -1,7 +1,8 @@
 use crate::behavior::persistent_parameters::NetworkingParametersRegistry;
-use crate::behavior::{Behavior, Event};
+use crate::behavior::{provider_storage, Behavior, Event};
 use crate::create::{
-    KADEMLIA_CONCURRENT_TASKS_BOOST_PER_PEER, REGULAR_CONCURRENT_TASKS_BOOST_PER_PEER,
+    ProviderOnlyRecordStore, KADEMLIA_CONCURRENT_TASKS_BOOST_PER_PEER,
+    REGULAR_CONCURRENT_TASKS_BOOST_PER_PEER,
 };
 use crate::request_responses::{Event as RequestResponseEvent, IfDisconnected};
 use crate::shared::{Command, CreatedSubscription, Shared};
@@ -13,6 +14,7 @@ use futures::{FutureExt, StreamExt};
 use libp2p::core::ConnectedPoint;
 use libp2p::gossipsub::{GossipsubEvent, TopicHash};
 use libp2p::identify::Event as IdentifyEvent;
+use libp2p::kad::store::RecordStore;
 use libp2p::kad::{
     AddProviderError, AddProviderOk, GetClosestPeersError, GetClosestPeersOk, GetProvidersError,
     GetProvidersOk, GetRecordError, GetRecordOk, InboundRequest, Kademlia, KademliaEvent,
@@ -73,14 +75,14 @@ enum QueryResultSender {
 
 /// Runner for the Node.
 #[must_use = "Node does not function properly unless its runner is driven forward"]
-pub struct NodeRunner<RecordStore>
+pub struct NodeRunner<ProviderStorage>
 where
-    RecordStore: Send + Sync + libp2p::kad::store::RecordStore + 'static,
+    ProviderStorage: provider_storage::ProviderStorage + Send + Sync + 'static,
 {
     /// Should non-global addresses be added to the DHT?
     allow_non_global_addresses_in_dht: bool,
     command_receiver: mpsc::Receiver<Command>,
-    swarm: Swarm<Behavior<RecordStore>>,
+    swarm: Swarm<Behavior<ProviderOnlyRecordStore<ProviderStorage>>>,
     shared_weak: Weak<Shared>,
     /// How frequently should random queries be done using Kademlia DHT to populate routing table.
     next_random_query_interval: Duration,
@@ -111,13 +113,13 @@ where
 }
 
 // Helper struct for NodeRunner configuration (clippy requirement).
-pub(crate) struct NodeRunnerConfig<RecordStore>
+pub(crate) struct NodeRunnerConfig<ProviderStorage>
 where
-    RecordStore: Send + Sync + libp2p::kad::store::RecordStore + 'static,
+    ProviderStorage: provider_storage::ProviderStorage + Send + Sync + 'static,
 {
     pub allow_non_global_addresses_in_dht: bool,
     pub command_receiver: mpsc::Receiver<Command>,
-    pub swarm: Swarm<Behavior<RecordStore>>,
+    pub swarm: Swarm<Behavior<ProviderOnlyRecordStore<ProviderStorage>>>,
     pub shared_weak: Weak<Shared>,
     pub next_random_query_interval: Duration,
     pub networking_parameters_registry: Box<dyn NetworkingParametersRegistry>,
@@ -127,12 +129,12 @@ where
     pub metrics: Option<Metrics>,
 }
 
-impl<RecordStore> NodeRunner<RecordStore>
+impl<ProviderStorage> NodeRunner<ProviderStorage>
 where
-    RecordStore: Send + Sync + libp2p::kad::store::RecordStore + 'static,
+    ProviderStorage: provider_storage::ProviderStorage + Send + Sync + 'static,
 {
     pub(crate) fn new(
-        NodeRunnerConfig::<RecordStore> {
+        NodeRunnerConfig {
             allow_non_global_addresses_in_dht,
             command_receiver,
             swarm,
@@ -143,7 +145,7 @@ where
             max_established_incoming_connections,
             max_established_outgoing_connections,
             metrics,
-        }: NodeRunnerConfig<RecordStore>,
+        }: NodeRunnerConfig<ProviderStorage>,
     ) -> Self {
         let (provider_records_sender, provider_records_receiver) =
             mpsc::channel(PROVIDER_RECORD_BUFFER_SIZE);
@@ -792,7 +794,7 @@ where
 
     // Returns `true` if query was cancelled
     fn unbounded_send_and_cancel_on_error<T>(
-        kademlia: &mut Kademlia<RecordStore>,
+        kademlia: &mut Kademlia<ProviderOnlyRecordStore<ProviderStorage>>,
         sender: &mut mpsc::UnboundedSender<T>,
         value: T,
         channel: &'static str,
